requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiolimiter==1.2.1
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
appdirs==1.4.4
arxiv==2.1.3
astunparse==1.6.3
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.2.0
Authlib==1.6.0
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
azure-identity==1.23.0
backoff==2.2.1
baidusearch==1.0.3
bcrypt==4.3.0
beartype==0.12.0
beautifulsoup4==4.13.3
bibtexparser==1.4.3
blinker==1.9.0
blis==1.0.1
boto3==1.37.22
botocore==1.37.22
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
cryptography==45.0.4
curl_cffi==0.11.3
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.6.7
datasets==3.2.0
defusedxml==0.7.1
diffusers==0.31.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docx2pdf==0.1.8
duckduckgo_search==7.5.5
english-words==2.0.1
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
Faker==37.1.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastmcp==2.8.0
feedparser==6.0.11
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googlesearch-python==1.3.0
greenlet==3.1.1
grpcio==1.68.0
gymnasium==1.0.0
h11==0.14.0
h5py==3.12.1
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
keyring==25.6.0
kiwisolver==1.4.7
langchain==0.3.23
langchain-anthropic==0.3.3
langchain-core==0.3.64
langchain-mcp-adapters==0.1.7
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
libvisualwebarena==0.0.15
libwebarena==0.0.4
loguru==0.7.3
lxml==5.3.1
magika==0.6.2
mammoth==1.9.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.14.1
markitdown==0.1.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
mcp==1.9.4
mdurl==0.1.2
mindmap-mcp-server==0.1.1
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msoffcrypto-tool==5.4.2
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.3.0
namex==0.0.8
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
olefile==0.47
ollama==0.4.7
onnxruntime==1.22.0
openai==1.66.5
openapi-pydantic==0.5.1
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.16
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pathlib==1.0.1
patsy==1.0.1
pdfminer.six==20250506
peewee==3.18.1
pillow==10.4.0
platformdirs==4.3.8
plotly==5.24.1
pluggy==1.5.0
portalocker==3.1.1
posthog==3.23.0
preshed==3.0.9
primp==0.14.0
propcache==0.2.0
Protego==0.4.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.13.1
PyMuPDF==1.26.1
pymupdf4llm==0.0.24
PyMySQL==1.1.1
PyNaCl==1.5.0
pypandoc==1.15
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pyreadline3==3.5.4
pytest==8.3.5
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzotero==1.6.11
readabilipy==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
s3transfer==0.11.4
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.0
seaborn==0.13.2
semanticscholar==0.8.4
setuptools==78.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
#spacy==3.8.2
#spacy-legacy==3.0.12
#spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.40
srsly==2.4.8
sse-starlette==2.2.1
starlette==0.46.1
statsmodels==0.14.4
streamlit==1.43.2
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.7.6
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
text-generation==0.7.0
thinc==8.3.2
thingspanel-mcp==0.1.2
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.0.20250328
types-tqdm==4.67.0.20250319
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2024.2
unidiff==0.7.5
urllib3==2.2.3
uvicorn==0.34.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webencodings==0.5.1
weblinx==0.3.2
weblinx-browsergym==0.0.1.dev14
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.0
xlrd==2.0.1
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.18.0
yfinance==0.2.31
youtube-transcript-api==1.0.3
zipp==3.21.0
zstandard==0.23.0
pdfplumber==0.11.0
autoviz==0.1.0
pyautogui==0.9.58
pybreaker==0.1.0